Update classpath settings: avoid duplicate build path for libs
When computing the classpath entry for a library found in the PDE
plugin, the result has export=true and no source attachment.
If the already existing classpath entry differs in whatever such a
parameter, the computed classpath entry is added as well, and Update
classpath settings fails with "Build path contains duplicate entry".

Instead, use the library path only to check for the already existing
classpath entry.